refactor(routes): change startupe route and home page name
Moved unsigned route to the startup route, and change home page route name
